refactor(onboard): simplify helper validation (#1885)
